+++ packages/components/atoms/f-button/src/components/Link.vue
@@ -3,3 +3,9 @@
         <slot />
     </a>
 </template>
+
+<script>
+export default {
+    name: 'LinkButton'
+};
+</script>