+++ packages/components/molecules/f-card-with-content/test/visual/f-card-with-content.visual.desktop.spec.js
@@ -1,5 +1,3 @@
-/* global browser */
-
 const CardWithContent = require('../../test-utils/component-objects/f-card-with-content.component');
 
 let cardWithContent = new CardWithContent();