+++ packages/components/molecules/f-navigation-links/test/accessibility/axe-accessibility.spec.js
@@ -1,4 +1,4 @@
-import { getAccessibilityTestResults } from '../../../../../../test/utils/axe-helper';
+import { getAccessibilityTestResults } from '../../../../../../test/utils/axe-helper'; // eslint-disable-line import/no-relative-packages
 
 const NavigationLinks = require('../../test-utils/component-objects/f-navigation-links.component');