+++ packages/components/molecules/f-searchbox/src/components/formElements/FormFullAddressSearchSuggestions.vue
@@ -10,6 +10,7 @@
                 $style.item,
                 { [$style.selected]: index === selected }
             ]"
+            type="button"
             @click="getSelectedStreetAddress($event, index, selected)">
             <map-pin-icon :class="$style['c-locationIcon']" />