+++ packages/components/organisms/f-content-cards/src/components/cardTemplates/StampCardPromotionCard.vue
@@ -188,5 +188,4 @@ $stampCard-promo-shadow: 0 2px 2px 0 rgba(0, 0, 0, 0.03),
         color: darken($color-content-link, $color-active-01);
     }
 }
-
 </style>