+++ packages/components/atoms/f-spinner/CHANGELOG.md
@@ -3,6 +3,13 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+Latest (to be added to next release)
+------------------------------
+*February 4, 2022*
+
+### Changed
+- Upgraded to ESLint v8
+
 
 v0.3.0
 ------------------------------